+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.1.10] - 5th May 2020
+
+* Control node movement on x and y axis via nodeMoveX, nodeMoveY
+* rename property size to nodeSize
+
 ## [0.1.8] - 5th May 2020
 
 * Sorted out vertical and horizontal stacking